Java/problems/_94_BinaryTreeInorderTraversal/Solution.java
package problems._94_BinaryTreeInorderTraversal;

import java.util.ArrayList;
import java.util.List;

class Solution {
    public List<Integer> inorderTraversal(TreeNode root) {
        List<Integer> values = new ArrayList<>();
        if (root == null) {
            return values;
        }

        values.addAll(new Solution().inorderTraversal(root.left));
        values.add(root.val);
        values.addAll(new Solution().inorderTraversal(root.right));

        return values;
    }

    public static class TreeNode {
        int val;
        TreeNode left;
        TreeNode right;

        TreeNode() {}

        TreeNode(int val) {
            this.val = val;
        }

        TreeNode(int val, TreeNode left, TreeNode right) {
            this.val = val;
            this.left = left;
            this.right = right;
        }
    }
}
